third_party/WebKit/LayoutTests/http/tests/security/xss-DENIED-getSVGDocument-iframe-expected.txt

CONSOLE MESSAGE: line 14: PASS: SVGDocument access threw: 'SecurityError: Failed to execute 'getSVGDocument' on 'HTMLIFrameElement': Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.'.
